Moved availableUpdates.go unit tests to a new file
Because it should be tracked in a different file and named as per parent
go file name. Also the cvo_test.go is pretty big now.
